Initial rework of moving program code to main memory
This is the initial pass of moving the program code into main system memory. As of yet I am not sure if I will go with this method.

None of the tests are converted over.

As it stands it is slower than the original version but it does allow some interesting things - such as self-modifying code.

A new memory region flag has been added: executable. This will indicate if a given section of memory can be used to form an opcode instruction or argument. Attempting to execute code from a non-executable region will throw an exception.

The concept of CPU memory block swapping has been introduced. By default a CPU can read code within it's allocated memory block, however it can be configured to allow it to read and execute code in any valid executable region.