Fix memory region overlap issue
Fix an issue where an item is requested from memory when it overlaps two memory regions.

Originally the code would return the last access flag entry from the list. However this is incorrect behaviour.

In this instance a list is made of any overlapping regions that match and the one with the lowest access permissions is selected. This will prevent previleage escallation.

Fixed an off-by-one error with memory region mapping.

The root memory region will now auto-resize as more memory regions are mapped.

All tests pass.