Memory access tests, memory type access methods and memory permission fix
Added a set of tests for memory access and for the new memory data type read/write methods.

Rework memory access permissions again as they were horribly broken before. In the case where a requested memory range crosses a memory region boundary then the region with the -highest- required permissions will be required, not the lowest.

This will mean that if a range crosses a public read/write and a private read/write region then the permissions of the latter will need to be met in order for access to be granted. Otherwise it would be possible to write into a higher permission memory region with insufficient privileges.

If an executable memory region is requested then -all- matched regions -must- have the executable flag set. The logic being the same as the permission logic explained above.

All tests pass.