Fix label argument binding code
Rewrite and hook up the label binding code and add tests for it.

Labels can now be bound to instruction arguments that are listed to support it. This will make working with jumps easier as the locations do not have to be hand-calculated going forward.

Fix the CPU step code since that has been used for the label tests.

Add early label replacement optimisation, closes #3.

All tests pass.