Rework the syntax for memory address references
Rework the syntax for memory address references, which are now specified by a prefix amperstand (&) instead of by square brackets (which are used to represent expressions) or by the a "$" (which is used to represent a hex literal value).

This will allow them to be more easier differentiated during parsing.

All tests pass.