Basic ASM parser implementation
A basic ASM parser implementation. No tests and still requires clean up.

All instructions now have a ArgumentRefTypes property that indicates the reference "type" of data held by the parameter for the instruction.
For example ADD_LIT_REG takes two parameters, a literal and a register so the reference types are "Literal Integer" and "Register" respectively. This data will be used in conjunction with the number and data type of the arguments to determine which opcode is a match for the parsed arguments of an instruction as quite a few share the same mnemonic form.

The syntax of the instructions has been slightly changed as the result of the parser work.
All "pointer" type reference types are prefixed with an "&".
All literals are prefixed with a "$" and their format can be specified with the following prefixed:
* 0b (binary)
* 0x (hexadecimal)
* 0 (octal)
* no prefix indicates a decimal number

Values from the disassembler can be set to output in hexadecimal or decimal by setting the "OutputLiteralsAsHex" property on a specific instruction, or on the "Instruction" base class to set it globally.

The expression parser has not been updated to support the new syntax in this push.