Optimise ASM parser with instruction data InsCacheEntry
Optimise the ASM parser by including an instruction data cache.

When the parser is constructed a dictionary of InsCacheEntry objects to opcode is constructed. Each known instruction is added to the dictionary.
As the instruction data is parsed a new InsCacheEntry is built and is checked against the list, returning the opcode if present. If not the instruction was invalid.

This should be faster than the old method and the code is also cleaner after this change.

All tests pass.