Resolve issue whereby the stack type hint cache falls out of sync with the contents of the stack
While working on the subroutine call code it became evident that the stack type hint cache falls out of sync with the actual contents of the stack.

This is because the CPU direclty sets the memory StackPointer to a different value which does not trigger an update to the type hint cache.

To resolve this the StackPointer Memory property is now private and cannot be set direclty outside of Memory. A new method (SetStackPointer) has been added for use outside of the Memory class and will enforce that the cache is correctly updated.

The SetDebuggingEnabled method has been removed from Memory as the type hint cache is now required, as explained above.

All tests pass.