Correctly handle stack and frame pointers
The original implementation of stack and frame pointers was broken and it was impossible to run nested subroutine calls.

The new code, thankfully, is also cleaner and simpler to understand. Hopefully.

The SetStackPointer method has been changed and no longer works with the stack type hint cache. This is down to the individual methods within the Memory class to handle. It now handles informing the CPU when the stack pointer changes, ensuring that the internal Stack Pointer register is always equal to the internal stack pointer property.

The CPU now no longer calls the SetStackPointer as a result of this and so the method has been marked as private within the Memory class.