Add tests for CALL_LIT and RET, bugfixes
Added tests for CALL_LIT and RET. I'm now relatively happy that they are covered in case of regressions.

Fixed a bug with the ASM parser that prevented it from matching call &$123 type instructions. This was resulting from the fact that when no labels were passed the checks were rejecting it on the basis that it could have a label supplied.
The test was faulty and has been replaced with a method that yields the correct results.

Fixed two bugs with the disassembly call subroutine label substitution system.
1. The system did not take account of the fact that a register can be passed as an argument.
2. The system did not take account of the fact that an invalid address passed as an argument. This would not necessarily execute correctly, but would raise an exception when attempting to disassemble the instruction.