Initial work on compiler changes
Initial work on compiler changes. Virtual machine will now load the binaries, though there is a lot of cleaning up and fixing to be done yet.

As it stands the debugger is running into issues because it reaches the end of the "code" section and starts to try to decode the "data" section. The best way around this is likely to be to move each useful binary section into it's own memory block. This will make it easier for the debugger to figure out where the end of a given section is plus it will make debugging easier.

Rather than loading the whole binary into memory, perhaps it is better to only load the parts that we can use. The headers, for example, are not useful for running the binary so they can be skipped.

Lots more work to be done.