Load binary files into fixed memory positions
Binary files now specify the position that they expect to be loaded (in memory). This makes calculating offsets simpler. A bunch of logic and code for offsetting has been removed from the instructions.

Added a new instruction argument type (InstructionSizeHint) for handling assembly like this: mov BYTE &R1, R2. This will now create an instruction that moves a byte of memory from the address pointed to by R1, rather than an int as would usually be retrieved.

Some experiments with a syntax highlighter for the disassembler. Unsure if I will pursue this yet.

Quick compiler doesn't work yet and unsure if I am going to fix it. It might just be removed instead.