Code clean up and adjustment to allocation of memory
Adjust the way that executable memory blocks are allocated within memory. If the entry point plus the length of the data is within the system memory region then it will allocate a section within the existing memory. Otherwise the memory will be expanded and then the segment added.

Some initial work in fixing all of the unit tests.